requirement.txt

click==7.1.2
configobj==5.0.6
constantly==15.1.0
cryptography==3.0
Django==3.1.5
django-admin-rangefilter==0.6.4
django-cors-headers==3.6.0
django-filter==2.4.0
django-import-export==2.5.0
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
jdcal==1.4.1
Markdown==3.3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.4.3
PyJWT==1.7.1
pyotp==2.6.0
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.2
xlrd==2.0.1
xlwt==1.3.0
